+++ server/usecases/user.test.ts
@@ -1,16 +1,31 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import type { EntitlementResult, QuotaEntitlementItem, UserAdminRepo, UserOperationFailure, UserWithOrg } from './ports'
+import type { Platform } from '../platform/interface'
+import type {
+  EntitlementResult,
+  ImageUpload,
+  ImageUploadResult,
+  ProfileRepo,
+  PublicUser,
+  QuotaEntitlementItem,
+  UserAdminRepo,
+  UserOperationFailure,
+  UserWithOrg,
+} from './ports'
 import {
+  type AvatarDeps,
   deleteUser,
   deleteUsers,
+  getPublicProfile,
   getUser,
   grantUserEntitlement,
   listUserEntitlements,
   listUsers,
+  removeAvatar,
   revokeUserEntitlement,
   setUserStatus,
   setUsersStatus,
   type UserDeps,
+  updateAvatar,
   updateUserEntitlement,
 } from './user'
 
@@ -364,5 +379,98 @@ describe('user usecase', () => {
       expect(out).toEqual({ ok: false, reason: 'not_found' })
       expect(record).not.toHaveBeenCalled()
     })
+  })
+})
+
+describe('avatar (self)', () => {
+  const AVATAR_PREFIX = '_system/avatars'
+  // platform is an opaque request-bound capability here — the usecase only
+  // forwards it to the gateway, so a sentinel is enough to assert pass-through.
+  const platform = { tag: 'platform' } as unknown as Platform
+  const sampleFile = new File([new Uint8Array(8)], 'a.png', { type: 'image/png' })
+
+  function makeAvatarDeps(image: Partial<ImageUpload> = {}) {
+    const setAvatar = vi.fn(async () => {})
+    const uploadPublicImage = vi.fn(async (): Promise<ImageUploadResult> => ({ ok: true, url: 'https://cdn/a.png' }))
+    const deletePublicImageVariants = vi.fn(async () => {})
+    const deps: AvatarDeps = {
+      imageUpload: { uploadPublicImage, deletePublicImageVariants, ...image } as ImageUpload,
+      profiles: { setAvatar } as unknown as ProfileRepo,
+    }
+    return { deps, setAvatar, uploadPublicImage, deletePublicImageVariants }
+  }
+
+  beforeEach(() => vi.clearAllMocks())
+
+  describe('updateAvatar', () => {
+    it('uploads, persists the url via setAvatar, and returns it', async () => {
+      const uploadPublicImage = vi.fn(
+        async (): Promise<ImageUploadResult> => ({ ok: true, url: 'https://cdn/_system/avatars/u1.png' }),
+      )
+      const { deps, setAvatar } = makeAvatarDeps({ uploadPublicImage })
+
+      const out = await updateAvatar(deps, { platform, userId: 'u1', file: sampleFile })
+
+      expect(out).toEqual({ ok: true, url: 'https://cdn/_system/avatars/u1.png' })
+      expect(uploadPublicImage).toHaveBeenCalledWith(platform, AVATAR_PREFIX, 'u1', sampleFile)
+      expect(setAvatar).toHaveBeenCalledWith('u1', 'https://cdn/_system/avatars/u1.png')
+    })
+
+    // The gateway owns which status a rejection carries (400 bad mime, 413 too
+    // large, 503 no public storage); the usecase surfaces it verbatim.
+    it.each([
+      { ok: false, status: 400, error: 'unsupported mime' },
+      { ok: false, status: 413, error: 'too large' },
+      { ok: false, status: 503, error: 'no public storage' },
+    ] satisfies ImageUploadResult[])('surfaces gateway failure ($status) and does not persist', async (failure) => {
+      const uploadPublicImage = vi.fn(async (): Promise<ImageUploadResult> => failure)
+      const { deps, setAvatar } = makeAvatarDeps({ uploadPublicImage })
+
+      const out = await updateAvatar(deps, { platform, userId: 'u1', file: sampleFile })
+
+      expect(out).toEqual(failure)
+      expect(setAvatar).not.toHaveBeenCalled()
+    })
+  })
+
+  describe('removeAvatar', () => {
+    it('clears the avatar in DB first, then best-effort removes storage variants', async () => {
+      const calls: string[] = []
+      const setAvatar = vi.fn(async () => {
+        calls.push('setAvatar')
+      })
+      const deletePublicImageVariants = vi.fn(async () => {
+        calls.push('deleteVariants')
+      })
+      const { deps } = makeAvatarDeps({ deletePublicImageVariants })
+      deps.profiles = { setAvatar } as unknown as ProfileRepo
+
+      await removeAvatar(deps, { platform, userId: 'u1' })
+
+      expect(setAvatar).toHaveBeenCalledWith('u1', null)
+      expect(deletePublicImageVariants).toHaveBeenCalledWith(platform, AVATAR_PREFIX, 'u1')
+      expect(calls).toEqual(['setAvatar', 'deleteVariants'])
+    })
+  })
+})
+
+describe('public profile', () => {
+  const samplePublicUser: PublicUser = { username: 'bob', name: 'Bob', image: null }
+  const withUser = (user: PublicUser | null) => ({
+    profiles: { getUserByUsername: async () => user, setAvatar: async () => {} } as ProfileRepo,
+  })
+
+  it('returns the public user', async () => {
+    expect(await getPublicProfile(withUser(samplePublicUser), 'bob')).toEqual(samplePublicUser)
+  })
+
+  it('returns null when the user does not exist', async () => {
+    expect(await getPublicProfile(withUser(null), 'ghost')).toBeNull()
+  })
+
+  it('queries by the given username', async () => {
+    const getUserByUsername = vi.fn(async () => samplePublicUser)
+    await getPublicProfile({ profiles: { getUserByUsername, setAvatar: async () => {} } as ProfileRepo }, 'alice')
+    expect(getUserByUsername).toHaveBeenCalledWith('alice')
   })
 })
